refactor: minimize pydantic ai v1 changes
